Graph: Add function to visualize stock price change as graph (#15)
* Graph Plotting function

This function intakes a filename of a specific stock *.log then it will plot a graph with the recent 71 data. Item stockA.log is an example of how this function intake data. The rising trend is shown in green while the dropping trend is shown in red.

* Complete Graph Plotting Function

This function intakes a stock name, the stock price history should be stored in *.log. stockA.log is given as reference. It will the take the most recent 71 data points to plot a stock price history graph

* Graph Plotting Function

* formatting

* formatting

* bug fixed for graph plotting

* minor fixes